Cleaning Up
Breaking the slf4j log implementation into its own class to make
everything easier to read and then cleaning up the slf4j interface to
follow basic java best practices.
